// src/app/modules/comercial/gestao/ranking-clientes/models/ranking.ts
export interface IRanking{
    codClassificacao: number;
    nomeClassificacao: string;
    peso: number;
    nota1Fim: number;
    nota2Inicio: number;
    nota2Fim: number;
    nota3Inicio: number;
    nota3Fim: number;
    nota4Inicio: number;
    nota4Fim: number;
    nota5Inicio: number;
    situacao: number;
}